requirements.txt
flask
openai
python-docx
requests
